Updating api to use doubles

+++ envi-api/src/main/java/com/envi/api/enviapi/model/Data.java
@@ -15,8 +15,8 @@ public class Data {
     @Id @GeneratedValue private Long id;
     private String name;
     private String description;
-    private Long longitude;
-    private Long latitude;
+    private Double longitude;
+    private Double latitude;
     private String author;
     @OneToMany(fetch=FetchType.EAGER)
     private List<Comments> commentsList;
